remove auth from ADO console comment creator, update tests
