+++ packages/ado-extension/src/progress-reporter/console/ado-console-comment-creator.spec.ts
@@ -1,12 +1,15 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 import 'reflect-metadata';
+
 import * as adoTask from 'azure-pipelines-task-lib/task';
 import * as nodeApi from 'azure-devops-node-api';
 import { Mock, Times, IMock, MockBehavior } from 'typemoq';
-import { ADOTaskConfig } from '../../task-config/ado-task-config';
-import { Logger, ReportMarkdownConvertor } from '@accessibility-insights-action/shared';
 import { AdoConsoleCommentCreator } from './ado-console-comment-creator';
+import { ADOTaskConfig } from '../../task-config/ado-task-config';
+import { CombinedReportParameters } from 'accessibility-insights-report';
+
+import { Logger, ReportMarkdownConvertor } from '@accessibility-insights-action/shared';
 
 describe(AdoConsoleCommentCreator, () => {
     let adoTaskMock: IMock<typeof adoTask>;
@@ -15,15 +18,7 @@ describe(AdoConsoleCommentCreator, () => {
     let nodeApiMock: IMock<typeof nodeApi>;
     let reportMarkdownConvertorMock: IMock<ReportMarkdownConvertor>;
     let webApiMock: IMock<nodeApi.WebApi>;
-    let consoleCommentCreator: AdoConsoleCommentCreator;
-
-    const handlerStub = {
-        prepareRequest: () => {
-            return;
-        },
-        canHandleAuthentication: () => false,
-        handleAuthentication: () => Promise.reject(),
-    };
+    let adoConsoleCommentCreator: AdoConsoleCommentCreator;
 
     beforeEach(() => {
         adoTaskMock = Mock.ofType<typeof adoTask>(undefined, MockBehavior.Strict);
@@ -35,46 +30,41 @@ describe(AdoConsoleCommentCreator, () => {
     });
 
     describe('constructor', () => {
-        it('should not initialize if missing required variable', () => {
-            setupInitializeWithTokenServiceConnection();
-            setupInitializeMissingVariable();
-
-            expect(() => buildConsoleCreatorWithMocks()).toThrow('Unable to find System.TeamFoundationCollectionUri');
+        it('should initialize', () => {
+            adoConsoleCommentCreator = buildAdoConsoleCommentCreatorWithMocks();
 
             verifyAllMocks();
         });
+    });
 
-        it('should not initialize if serviceConnection uses unsupported auth', () => {
-            setupInitializeWithUnsupportedServiceConnection();
+    describe('completeRun', () => {
+        it('should output results to the console', async () => {
+            const reportStub: CombinedReportParameters = {
+                results: {
+                    urlResults: {
+                        failedUrls: 1,
+                    },
+                },
+            } as CombinedReportParameters;
+            const baselineInfoStub = {};
+            const reportMarkdownStub = '#ReportMarkdownStub';
 
-            expect(() => buildConsoleCreatorWithMocks()).toThrow('Unsupported auth scheme. Please use token or basic auth.');
+            const expectedLogOutput = AdoConsoleCommentCreator.CURRENT_COMMENT_TITLE + reportMarkdownStub;
 
-            verifyAllMocks();
-        });
+            adoTaskConfigMock
+                .setup((atcm) => atcm.getBaselineFile())
+                .returns(() => undefined)
+                .verifiable(Times.once());
 
-        it('should initialize if isSupported returns true and serviceConnection uses basic auth', () => {
-            setupInitializeWithBasicServiceConnection();
-            setupInitializeSetConnection(webApiMock.object);
+            reportMarkdownConvertorMock
+                .setup((o) => o.convert(reportStub, AdoConsoleCommentCreator.CURRENT_COMMENT_TITLE, baselineInfoStub))
+                .returns(() => expectedLogOutput)
+                .verifiable(Times.once());
 
-            consoleCommentCreator = buildConsoleCreatorWithMocks();
+            loggerMock.setup((lm) => lm.logInfo(expectedLogOutput)).verifiable(Times.once());
 
-            verifyAllMocks();
-        });
-
-        it('should initialize if isSupported returns true and serviceConnection uses token auth', () => {
-            setupInitializeWithTokenServiceConnection();
-            setupInitializeSetConnection(webApiMock.object);
-
-            consoleCommentCreator = buildConsoleCreatorWithMocks();
-
-            verifyAllMocks();
-        });
-
-        it('should initialize if isSupported returns true and serviceConnectionName is not set', () => {
-            setupInitializeWithoutServiceConnectionName();
-            setupInitializeSetConnection(webApiMock.object);
-
-            consoleCommentCreator = buildConsoleCreatorWithMocks();
+            adoConsoleCommentCreator = buildAdoConsoleCommentCreatorWithMocks();
+            await adoConsoleCommentCreator.completeRun(reportStub);
 
             verifyAllMocks();
         });
@@ -83,17 +73,15 @@ describe(AdoConsoleCommentCreator, () => {
     describe('failRun', () => {
         it('does nothing interesting', async () => {
             const message = 'message';
-            setupInitializeWithoutServiceConnectionName();
-            setupInitializeSetConnection(webApiMock.object);
-            consoleCommentCreator = buildConsoleCreatorWithMocks();
+            adoConsoleCommentCreator = buildAdoConsoleCommentCreatorWithMocks();
 
-            await consoleCommentCreator.failRun(message);
+            await adoConsoleCommentCreator.failRun(message);
 
             verifyAllMocks();
         });
     });
 
-    const buildConsoleCreatorWithMocks = () =>
+    const buildAdoConsoleCommentCreatorWithMocks = () =>
         new AdoConsoleCommentCreator(
             adoTaskConfigMock.object,
             reportMarkdownConvertorMock.object,
@@ -108,132 +96,5 @@ describe(AdoConsoleCommentCreator, () => {
         loggerMock.verifyAll();
         reportMarkdownConvertorMock.verifyAll();
         webApiMock.verifyAll();
-    };
-
-    const setupBaselineFileParameterDoesNotExist = () => {
-        adoTaskConfigMock
-            .setup((o) => o.getBaselineFile())
-            .returns(() => undefined)
-            .verifiable(Times.atLeastOnce());
-    };
-
-    const setupInitializeWithoutServiceConnectionName = () => {
-        const apitoken = 'token';
-
-        adoTaskConfigMock
-            .setup((o) => o.getRepoServiceConnectionName())
-            .returns(() => '')
-            .verifiable(Times.once());
-        adoTaskMock
-            .setup((o) => o.getEndpointAuthorizationParameter('SystemVssConnection', 'AccessToken', false))
-            .returns(() => apitoken)
-            .verifiable(Times.once());
-        nodeApiMock
-            .setup((o) => o.getPersonalAccessTokenHandler(apitoken))
-            .returns(() => handlerStub)
-            .verifiable(Times.once());
-    };
-
-    const setupInitializeWithTokenServiceConnection = () => {
-        const apitoken = 'token';
-
-        const serviceConnection = 'service-connection';
-        const endpointAuthorizationStub: adoTask.EndpointAuthorization = {
-            parameters: {
-                apitoken,
-            },
-            scheme: '',
-        };
-
-        adoTaskConfigMock
-            .setup((o) => o.getRepoServiceConnectionName())
-            .returns(() => serviceConnection)
-            .verifiable(Times.once());
-        adoTaskMock
-            .setup((o) => o.getEndpointAuthorization(serviceConnection, false))
-            .returns(() => endpointAuthorizationStub)
-            .verifiable(Times.once());
-        adoTaskMock
-            .setup((o) => o.getEndpointAuthorizationScheme(serviceConnection, true))
-            .returns(() => 'Token')
-            .verifiable(Times.once());
-        nodeApiMock
-            .setup((o) => o.getPersonalAccessTokenHandler(apitoken))
-            .returns(() => handlerStub)
-            .verifiable(Times.once());
-    };
-
-    const setupInitializeWithBasicServiceConnection = () => {
-        const serviceConnection = 'service-connection',
-            //[SuppressMessage("Microsoft.Security", "CS002:SecretInNextLine", Justification="Fake creds")]
-            username = 'user',
-            //[SuppressMessage("Microsoft.Security", "CS002:SecretInNextLine", Justification="Fake creds")]
-            password = 'secret';
-        const endpointAuthorizationStub: adoTask.EndpointAuthorization = {
-            parameters: {
-                username,
-                password,
-            },
-            scheme: '',
-        };
-
-        adoTaskConfigMock
-            .setup((o) => o.getRepoServiceConnectionName())
-            .returns(() => serviceConnection)
-            .verifiable(Times.once());
-        adoTaskMock
-            .setup((o) => o.getEndpointAuthorization(serviceConnection, false))
-            .returns(() => endpointAuthorizationStub)
-            .verifiable(Times.once());
-        adoTaskMock
-            .setup((o) => o.getEndpointAuthorizationScheme(serviceConnection, true))
-            .returns(() => 'UsernamePassword')
-            .verifiable(Times.once());
-
-        nodeApiMock
-            .setup((o) => o.getBasicHandler(username, password))
-            .returns(() => handlerStub)
-            .verifiable(Times.once());
-    };
-
-    const setupInitializeWithUnsupportedServiceConnection = () => {
-        const serviceConnection = 'service-connection';
-        const endpointAuthorizationStub: adoTask.EndpointAuthorization = {
-            parameters: {},
-            scheme: '',
-        };
-
-        adoTaskConfigMock
-            .setup((o) => o.getRepoServiceConnectionName())
-            .returns(() => serviceConnection)
-            .verifiable(Times.once());
-        adoTaskMock
-            .setup((o) => o.getEndpointAuthorization(serviceConnection, false))
-            .returns(() => endpointAuthorizationStub)
-            .verifiable(Times.once());
-        adoTaskMock
-            .setup((o) => o.getEndpointAuthorizationScheme(serviceConnection, true))
-            .returns(() => undefined)
-            .verifiable(Times.once());
-    };
-
-    const setupInitializeSetConnection = (connection: nodeApi.WebApi) => {
-        const url = 'url';
-
-        adoTaskMock
-            .setup((o) => o.getVariable('System.TeamFoundationCollectionUri'))
-            .returns(() => url)
-            .verifiable(Times.atLeastOnce());
-        nodeApiMock
-            .setup((o) => new o.WebApi(url, handlerStub))
-            .returns(() => connection)
-            .verifiable(Times.once());
-    };
-
-    const setupInitializeMissingVariable = () => {
-        adoTaskMock
-            .setup((o) => o.getVariable('System.TeamFoundationCollectionUri'))
-            .returns(() => undefined)
-            .verifiable(Times.once());
     };
 });
